magicbox-node/workers/base/requirements.txt
# Base requirements for MagicBox workers
nats-py>=2.6.0
aiohttp>=3.9.0
numpy>=1.24.0

# Optional: for frame decoding
# opencv-python>=4.8.0

